+++ Sources/TABResourceLoader/Protocols/Resource types/JSONArrayResourceType.swift
@@ -17,14 +17,14 @@ public protocol JSONArrayResourceType: DataResourceType {
 
    - parameter jsonArray: The JSON array to parse
 
-   - returns: An instantiated model if parsing was succesful, otherwise nil
+   - returns: An instantiated model if parsing was successful, otherwise throws
    */
   func model(from jsonArray: [Any]) throws -> Model
 }
 
 extension JSONArrayResourceType {
 
-  public func result(from data: Data) throws -> Model {
+  public func model(from data: Data) throws -> Model {
     guard let jsonObject = try? JSONSerialization.jsonObject(with: data, options: .mutableContainers) else {
       throw JSONParsingError.invalidJSONData
     }
@@ -32,11 +32,8 @@ extension JSONArrayResourceType {
     guard let jsonArray = jsonObject as? [Any] else {
       throw JSONParsingError.notAJSONArray
     }
-    do {
-      return try model(from: jsonArray)
-    } catch {
-      throw error
-    }
+
+    return try model(from: jsonArray)
   }
 
 }