+++ Tests/TABResourceLoaderTests/Protocols/JSONDictionaryResourceTypeTests.swift
@@ -14,7 +14,7 @@ class JSONDictionaryResourceTypeTests: XCTestCase {
   func test_invalidJSONData() {
     let mockJSONObjectResourceType = MockJSONDictionaryResourceType()
     do {
-      let _ = try mockJSONObjectResourceType.result(from: Data())
+      let _ = try mockJSONObjectResourceType.model(from: Data())
       XCTFail("No error found")
     } catch {
       XCTAssertEqual(error as? JSONParsingError, JSONParsingError.invalidJSONData)
@@ -26,7 +26,7 @@ class JSONDictionaryResourceTypeTests: XCTestCase {
     let jsonArray = ["invalid_key"]
     let data = serialize(jsonObject: jsonArray)
     do {
-      let _ = try mockJSONObjectResourceType.result(from: data)
+      let _ = try mockJSONObjectResourceType.model(from: data)
       XCTFail("No error found")
     } catch {
       XCTAssertEqual(error as? JSONParsingError, JSONParsingError.notAJSONDictionary)
@@ -38,7 +38,7 @@ class JSONDictionaryResourceTypeTests: XCTestCase {
     let jsonDictionary = ["invalid_key": "mock"]
     let data = serialize(jsonObject: jsonDictionary)
     do {
-      let _ = try mockJSONObjectResourceType.result(from: data)
+      let _ = try mockJSONObjectResourceType.model(from: data)
       XCTFail("No error found")
     } catch {
       XCTAssertEqual(error as? JSONParsingError, JSONParsingError.cannotParseJSONDictionary)
@@ -49,7 +49,7 @@ class JSONDictionaryResourceTypeTests: XCTestCase {
     let mockJSONObjectResourceType = MockJSONDictionaryResourceType()
     let jsonDictionary = ["name": "mock"]
     let data = serialize(jsonObject: jsonDictionary)
-    guard let calculatedMockObject = try? mockJSONObjectResourceType.result(from: data) else {
+    guard let calculatedMockObject = try? mockJSONObjectResourceType.model(from: data) else {
       XCTFail("No error found")
       return
     }